backend/notetaking/requirements.txt
# Core dependencies
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter

# Development tools (optional)
django-extensions>=3.2  # Extra management commands
python-dotenv>=1.0  # For managing environment variables
drf-spectacular 
drf-spectacular-sidecar
